make slicing explicit

+++ src/kernels.cpp
@@ -3,20 +3,20 @@
 
 void addKernel(Tensor &result, const Tensor &left, const Tensor &right) {
   for (auto mIdx : MultiIndexRange(result.shape())) {
-    result[mIdx] = static_cast<double>(left[mIdx]) + static_cast<double>(right[mIdx]);
+    result[mIdx] = left[mIdx] + right[mIdx];
   }
 }
 
 void multKernel(Tensor &result, const double scalar, const Tensor &tensor) {
   for (auto mIdx : MultiIndexRange(result.shape())) {
-    result[mIdx] = scalar * static_cast<double>(tensor[mIdx]);
+    result[mIdx] = scalar * tensor[mIdx];
   }
 }
 
 void multKernel(Tensor &result, const Tensor &left, const Tensor &right) {
   MultiIndex resultIdx(result.shape());
   for (size_t i = 0; i < left.size(); ++i) {
-    result[resultIdx] = static_cast<double>(left[resultIdx]) * static_cast<double>(right[resultIdx]);
+    result[resultIdx] = left[resultIdx] * right[resultIdx];
     ++resultIdx;
   }
 }
